Alterado post

+++ src/main/java/com/lucascosta/jaegertracingserver/model/Pessoa.java
@@ -2,11 +2,13 @@ package com.lucascosta.jaegertracingserver.model;
 
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 public class Pessoa {
     String nome = "Lucas";
     String sobreNome = "Dealis";
     Integer idade = 20;
     String regiao = "Teotonio";
-    Integer cpf = 123456789;
+    BigInteger cpf = BigInteger.valueOf(123456789);
 }
